+++ Wishlist/src/main/java/com/demo/wishlist/WishlistService.java
@@ -2,15 +2,21 @@ package com.demo.wishlist;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestOperations;
+import org.springframework.web.client.RestTemplate;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @AllArgsConstructor
 @Service
 public class WishlistService {
 
     WishlistRepository wishlistRepository;
+    private RestTemplate restTemplate;
+
+    // Add checks for administration and username.
 
 
     public void createWishlist(String title, String username){
@@ -29,26 +35,54 @@ public class WishlistService {
         return  wishlistRepository.findAllByUsername(username);
     }
 
-    public Wishlist addBook(int bookId, int wishlistId){
-        Wishlist w = wishlistRepository.findById(wishlistId).get();
-        w.addBook(bookId);
-        wishlistRepository.save(w);
-        return w;
+    public Wishlist addBook(int bookId, int wishlistId,String username){
+        // CHECK BOOK ID FOR VALID ??
+        Optional<Wishlist> optionalWishlist = wishlistRepository.findById(wishlistId);
+        if(optionalWishlist.isPresent()){
+            Wishlist w = optionalWishlist.get();
+            if(w.getUsername().equals(username) || isAdmin(username)){
+                w.addBook(bookId);
+                wishlistRepository.save(w);
+                return w;
+            }
+        }
+        return null;
     }
 
-    public void removeBook(int bookId, int wishListId){
-        Wishlist w = wishlistRepository.findById(wishListId).get();
-        w.removeBook(bookId);
-        wishlistRepository.save(w);
+    public void removeBook(int bookId, int wishListId, String username){
+        Optional<Wishlist> optionalWishlist = wishlistRepository.findById(wishListId);
+        if(optionalWishlist.isPresent()){
+            Wishlist w = optionalWishlist.get();
+            if(w.getUsername().equals(username) || isAdmin(username)){
+                w.removeBook(bookId);
+                wishlistRepository.save(w);
+            }
+        }
     }
 
-    public void removeWishlist(int wishListId){
-        wishlistRepository.deleteById(wishListId);
+    public void removeWishlist(int wishListId, String username){
+        Optional<Wishlist> optionalWishlist = wishlistRepository.findById(wishListId);
+        if(optionalWishlist.isPresent()){
+            Wishlist w = optionalWishlist.get();
+            if(w.getUsername().equals(username) || isAdmin(username)){
+                wishlistRepository.deleteById(wishListId);
+            }
+        }
     }
 
-    public void updateTitle(String title, int wishListId) {
-        Wishlist w = wishlistRepository.findById(wishListId).get();
-        w.setTitle(title);
-        wishlistRepository.save(w);
+    public void updateTitle(String title, int wishListId, String username) {
+        Optional<Wishlist> optionalWishlist = wishlistRepository.findById(wishListId);
+        if(optionalWishlist.isPresent()){
+            Wishlist w = optionalWishlist.get();
+            if(w.getUsername().equals(username) || isAdmin(username)){
+                w.setTitle(title);
+                wishlistRepository.save(w);
+            }
+        }
+    }
+    public boolean isAdmin(String username){
+        return Boolean.TRUE.equals(restTemplate.getForObject(
+                "lb://AUTHORIZATION/Authorization/IsAdmin/" + username
+                , Boolean.class));
     }
 }
